fix net loss bincount from int to float

+++ nets/SBMs_node_classification/gin_net.py
@@ -73,7 +73,7 @@ class GINNet(nn.Module):
 
         # calculating label weights for weighted loss computation
         V = label.size(0)
-        label_count = torch.bincount(label)
+        label_count = torch.bincount(label).float()
         label_count = label_count[label_count.nonzero()].squeeze()
         cluster_sizes = torch.zeros(self.n_classes).long().to(self.device)
         cluster_sizes[torch.unique(label)] = label_count
